Rename utils to setDarkmode in context
